Allow installation of system packages (required by some R packages)
For example in this case the shiny app would need to use the `rgl` R package.
Try to install this would fail because of the lack of some Ubuntu packages
(`libgl1-mesa-dev` and `libglu1-mesa-dev` in this case).

The `apt-dependencies` file will contains a list of debian packages separated
by spaces to be installed using `apt-get install ...` when the docker
image is built, before the R packages installed by packrat.