Use AP Shiny Server v0.0.5
This version has more dependencies upgraded and 2 more security fixes.
See PR for details: https://github.com/ministryofjustice/analytics-platform-shiny-server/pull/4

I've tested this locally with the Secure Geezer app and it seems to work as
before but my plan is to ask a couple of our users to try this new
version on some non-critical webapps to see if it cause any problems.

Part of ticket: https://trello.com/c/PLR5Dft2